Create ShoolAI.md
This is a collaborative problem-solving lab that I completed for my Machine Learning Fundamental course. For this lab, I partnered with a pre-trained chatbot to a have dialog that was mostly me being question. SchoolAI chatbot asked questions that challenged my understanding on Supervised Learning, Unsupervised Learning, Reinforcement Learning, the importance of Algorithms, the process for training AI models and Repetition in Training. I engaged in meaningful dialogue with a SchoolAI chatbot to solve complex problems. I used SchoolAI chatbot as guidance to connect theory to real-world machine learning.